requirements.txt
# PydanticAI und A2A Support
pydantic-ai-slim[a2a]
fasta2a

# Database Connectors
pymongo
sqlalchemy
psycopg2-binary
mysql-connector-python

# Data Processing
pandas
numpy

# Web Framework
fastapi
uvicorn

# Utilities
python-dotenv
logging
pathlib
